moving CCA and SSVEPfiltering in pipelines
