+++ examples/plot_cross_subject_ssvep.py
@@ -13,9 +13,7 @@ We will use the SSVEP paradigm, which uses the AUC as metric.
 # License: BSD (3-clause)
 
 from sklearn.pipeline import make_pipeline
-from sklearn.cross_decomposition import CCA
 from sklearn.linear_model import LogisticRegression
-from sklearn.base import BaseEstimator, TransformerMixin, ClassifierMixin
 
 from pyriemann.tangentspace import TangentSpace
 from pyriemann.estimation import Covariances
@@ -23,9 +21,9 @@ from pyriemann.estimation import Covariances
 from moabb.evaluations import CrossSubjectEvaluation
 from moabb.paradigms import SSVEP, FilterBankSSVEP
 from moabb.datasets import SSVEPExo
+from moabb.pipelines import SSVEP_CCA, ExtendedSSVEPSignal
 import moabb
 
-import numpy as np
 import pandas as pd
 import seaborn as sns
 import matplotlib.pyplot as plt
@@ -36,94 +34,15 @@ warnings.simplefilter(action='ignore', category=RuntimeWarning)
 moabb.set_log_level('info')
 
 ###############################################################################
-# Auxiliary functions
-# -------------------
+# Loading dataset
+# ---------------
 #
-# We will need an auxiliary transformer for filtering the signal.
-# This auxiliary transformer allows to reshape the signal for building
-# extended covariances matrices
-
-
-class FilteredSignal(BaseEstimator, TransformerMixin):
-
-    def __init__(self):
-        pass
-
-    def fit(self, X, y):
-        """fit."""
-        return self
-
-    def transform(self, X):
-        """transform. """
-        out = X.transpose((0, 3, 1, 2))
-        n_trials, n_freqs, n_channels, n_times = out.shape
-        out = out.reshape((n_trials, n_channels * n_freqs, n_times))
-        return out
-
-###############################################################################
-# The following class defines a SSVEP CCA classifier, where the CCA is computed
-# from the set of training signals and some pure sinusoids to act as reference.
-# Classification is made by taking the frequency with the max correlation, as
-# proposed in Bin et al. (2009) https://doi.org/10.1088/1741-2560/6/4/046002
-
-
-class SSVEP_CCA(BaseEstimator, ClassifierMixin):
-
-    def __init__(self, interval, freqs, n_harmonics=3):
-        self.Yf = dict()
-        self.cca = CCA(n_components=1)
-        self.slen = interval[1] - interval[0]
-        self.freqs = freqs
-        self.n_harmonics = n_harmonics
-        self.one_hot = {}
-        for i, k in enumerate(freqs.keys()):
-            self.one_hot[k] = i
-
-    def fit(self, X, y, sample_weight=None):
-        """fit."""
-        n_times = X.shape[2]
-
-        for f in self.freqs:
-            if f.replace('.', '', 1).isnumeric():
-                freq = float(f)
-                yf = []
-                for h in range(1, self.n_harmonics + 1):
-                    yf.append(np.sin(2 * np.pi * freq * h *
-                                     np.linspace(0, self.slen, n_times)))
-                    yf.append(np.cos(2 * np.pi * freq * h *
-                                     np.linspace(0, self.slen, n_times)))
-                self.Yf[f] = np.array(yf)
-        return self
-
-    def predict(self, X):
-        """predict"""
-        y = []
-        for i, x in enumerate(X):
-            corr_f = {}
-            for f in self.freqs:
-                if f.replace('.', '', 1).isnumeric():
-                    S_x, S_y = self.cca.fit_transform(x.T, self.Yf[f].T)
-                    corr_f[f] = np.corrcoef(S_x.T, S_y.T)[0, 1]
-            y.append(self.one_hot[max(corr_f, key=lambda k: corr_f[k])])
-        return y
-
-    def predict_proba(self, X):
-        """predict proba"""
-        P = np.zeros(shape=(len(X), len(self.freqs)))
-        for i, x in enumerate(X):
-            for j, f in enumerate(self.freqs):
-                if f.replace('.', '', 1).isnumeric():
-                    S_x, S_y = self.cca.fit_transform(x.T, self.Yf[f].T)
-                    P[i, j] = np.corrcoef(S_x.T, S_y.T)[0, 1]
-        return P / np.resize(P.sum(axis=1), P.T.shape).T
-
-
-# Loading the SSVEP paradigm and the SSVEP Exo dataset, restricting to
-# the first three classes (here 13, 17 and 21 Hz) and the first 2 subjects.
-# Two paradigms are created, one for the CCA classification and another one
-# based on filter-bank decomposition for designing extended covariance
-# matrices.
-
+# We will load the data from the first 2 subjects of the SSVEP_Exo dataset
+# and compare two algorithms on this set. One of the algorithm could only
+# process class associated with a stimulation frequency, we will thus drop
+# the resting class. As the resting class is the last defined class, picking
+# the first three classes (out of four) allows to focus only on the stimulation
+# frequency.
 
 n_subject = 2
 for i in range(n_subject):
@@ -153,7 +72,7 @@ freqs = paradigm.used_events(dataset)
 
 pipelines_fb = {}
 pipelines_fb['RG + LogReg'] = make_pipeline(
-    FilteredSignal(),
+    ExtendedSSVEPSignal(),
     Covariances(estimator='lwf'),
     TangentSpace(),
     LogisticRegression(solver='lbfgs', multi_class='auto'))
@@ -161,7 +80,6 @@ pipelines_fb['RG + LogReg'] = make_pipeline(
 pipelines = {}
 pipelines['CCA'] = make_pipeline(
     SSVEP_CCA(interval=interval, freqs=freqs, n_harmonics=3))
-#    BroadbandSignal(),
 
 ##############################################################################
 # Evaluation
